feat(server): use `common` pkg
