+++ internal/server/handlers.go
@@ -6,10 +6,10 @@ import (
 	"net/http"
 	"time"
 
+	auth "github.com/codespace-operator/common/auth/pkg/auth"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	codespacev1 "github.com/codespace-operator/codespace-operator/api/v1"
-	"github.com/codespace-operator/codespace-operator/internal/auth"
 	"github.com/codespace-operator/codespace-operator/internal/common"
 )
