+++ internal/server/helpers.go
@@ -8,11 +8,13 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/fs"
+	"log"
 	"net"
 	"net/http"
 	"net/url"
 	"sort"
 	"strings"
+	"time"
 
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -220,6 +222,21 @@ func discoverNamespacesWithSessions(ctx context.Context, deps *serverDeps) ([]st
 	return sessionNamespaces, nil
 }
 
+// testKubernetesConnection verifies that we can connect to the Kubernetes API
+func testKubernetesConnection(c client.Client) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var sessionList codespacev1.SessionList
+	// It would make more sense to test to its own namespace
+	ns, _ := common.ResolveAnchorNamespace()
+	if err := c.List(ctx, &sessionList, client.InNamespace(ns), client.Limit(1)); err != nil {
+		return fmt.Errorf("failed to connect to Kubernetes API: %w", err)
+	}
+	log.Printf("✅ Successfully connected to Kubernetes API (found %d sessions in %s namespace)", len(sessionList.Items), ns)
+	return nil
+}
+
 // getAllowedNamespacesForUser determines which namespaces a user can access
 // by checking their RBAC permissions against known namespaces
 func getAllowedNamespacesForUser(ctx context.Context, deps *serverDeps, subject string, roles []string) ([]string, error) {
@@ -239,7 +256,7 @@ func getAllowedNamespacesForUser(ctx context.Context, deps *serverDeps, subject 
 
 	// Check each namespace individually
 	for _, ns := range allNamespaces {
-		if canAccess, _ := deps.rbac.CanAccessNamespace(subject, roles, ns); canAccess {
+		if canAccess, _ := deps.rbac.Enforce(subject, roles, "session", "list", ns); canAccess {
 			allowedNamespaces = append(allowedNamespaces, ns)
 		}
 	}
